Separate Modal Component
State and various things should be controlled by the parent component
