+++ frontend/src/components/ClubBoard/ClubBoard.tsx
@@ -4,7 +4,10 @@ import Filter from '../Filter/Filter';
 import { ClubType } from '../../../../backend/types/types';
 import './ClubBoard.css';
 import NavBar from '../NavBar/NavBar';
-import UploadImageModal from '../UploadImageModal/UploadImageModal';
+import Modal from '../Modal/Modal';
+
+// FOR UPLOAD IMAGE MODAL
+import { xButton, imgIcon, cloud } from '../../icons/image-modal';
 
 const ClubBoard = () => {
   const [origClubs, setOrigClubs] = useState<ClubType[]>([]); // an array of all clubs
@@ -41,22 +44,173 @@ const ClubBoard = () => {
     );
   }
 
-  // TODO: temporary trigger to test the modal; needs to be removed later
+  // BEGIN: FOR UPLOAD IMAGE MODAL
   const [showUploadModal, setShowModal] = useState(false);
+
+  function getFileExt(fname: string) {
+    return fname.slice(((fname.lastIndexOf('.') - 1) >>> 0) + 2).toLowerCase();
+  }
+
+  function isAnImage(file: string) {
+    return 'jpg png jpeg gif'.includes(getFileExt(file));
+  }
+
+  const [selectedFile, setSelectedFile] = useState<File>();
+
+  let defaultMessage = '';
+  let tooManyFilesMsg = 'Only one file at a time please!';
+  let notAnImgMsg = "This file isn't a valid image!";
+  let notAFileMsg = "This isn't a valid file!";
+  const [message, setMessage] = useState(defaultMessage);
+
+  const [hoveringDrop, setHoveringDrop] = useState(false);
+
+  function updateFile(file: File | undefined) {
+    setSelectedFile(file);
+    setMessage(defaultMessage);
+    setHoveringDrop(false);
+  }
+
+  const handleDrop = (ev: any) => {
+    // Prevent default behavior (Prevent file from being opened)
+    ev.preventDefault();
+
+    if (ev.dataTransfer.items) {
+      // Use DataTransferItemList interface to access the file(s)
+      if (ev.dataTransfer.items.length > 1) {
+        setMessage(tooManyFilesMsg);
+        setHoveringDrop(false);
+      } else {
+        if (ev.dataTransfer.items[0].kind === 'file') {
+          var file = ev.dataTransfer.items[0].getAsFile();
+          if (isAnImage(file.name)) {
+            updateFile(file);
+          } else {
+            setMessage(notAnImgMsg);
+            setHoveringDrop(false);
+          }
+        } else {
+          setMessage(notAFileMsg);
+          setHoveringDrop(false);
+        }
+      }
+    } else {
+      // Use DataTransfer interface to access the file(s)
+      if (ev.dataTransfer.files.length > 1) {
+        setMessage(tooManyFilesMsg);
+        setHoveringDrop(false);
+      } else {
+        updateFile(ev.dataTransfer.files[0]);
+      }
+    }
+  };
+
+  const handleDragOver = (ev: any) => {
+    setHoveringDrop(true);
+
+    // setTimeout(function () {
+    //   setHoveringDrop(false);
+    // }, 5000); //turn it back to grey in 5 seconds
+
+    // Prevent default behavior (Prevent file from being opened)
+    ev.preventDefault();
+  };
+
+  const uploadImageFile = (file: File) => {
+    // TODO: UPLOAD THE FILE TO THE BACKEND
+    console.log(file.name);
+  };
+  // END: FOR UPLOAD IMAGE MODAL
 
   return (
     <>
       <NavBar hasSearch={true} callback={updateSearchText} />
 
-      {/* TODO: temporary components to test the modal; needs to be removed later */}
+      {/* TODO: temporary components to test the UPLOAD IMAGE MODAL; needs to be removed later */}
       <button onClick={() => setShowModal(true)}>Show Modal</button>
-      <UploadImageModal
+      <Modal
         show={showUploadModal}
-        onClose={() => setShowModal(false)}
-        // optionally:
-        sendFile={(file: File) => {
-          console.log(file.name);
+        onClose={() => {
+          setShowModal(false); // close modal
+          updateFile(undefined); // reset the file stored within the modal
         }}
+        onSubmit={() => {
+          if (selectedFile !== undefined) {
+            // if a file is selected
+            uploadImageFile(selectedFile); // send the file to backend to be uploaded
+          }
+        }}
+        title={'Upload Image'}
+        showSubmitButton={true}
+        bodyChildren={
+          <>
+            <div className="fileinfozone">
+              {selectedFile !== undefined ? (
+                <div className="fileinfo">
+                  <div className="fileNameAndIcon">
+                    <div className="fileIconWrapper">
+                      <img
+                        className="fileIcon"
+                        height="20"
+                        src={imgIcon}
+                        alt="Icon"
+                      />
+                    </div>
+                    <p className="filename">{selectedFile.name}</p>
+                  </div>
+                  <div className="fileIconWrapper">
+                    <img
+                      className="closeIcon"
+                      src={xButton}
+                      alt="Close the dialog"
+                      onClick={() => updateFile(undefined)}
+                    />
+                  </div>
+                </div>
+              ) : (
+                <></>
+              )}
+            </div>
+            <div className="dropzonewrapper">
+              <div
+                className={hoveringDrop ? 'dropzoneHovering' : 'dropzone'}
+                onDrop={handleDrop}
+                onDragOver={handleDragOver}
+              >
+                {
+                  // if the message is the default, that means we have no errors. display the default cloud icon:
+                  message === defaultMessage ? (
+                    <div className="infoCloud">
+                      <img
+                        className="cloudIcon"
+                        height="80"
+                        src={cloud}
+                        alt="Icon"
+                      />
+                      <br />
+                      Drag files here <br /> or
+                    </div>
+                  ) : (
+                    // otherwise, display an error message only
+                    <div className="errorMessage">{message}</div>
+                  )
+                }
+
+                <label htmlFor="hiddenFileInput" className="browseFiles">
+                  Browse Photos
+                </label>
+                <input
+                  className="hiddenFileInput"
+                  id="hiddenFileInput"
+                  type="file"
+                  name="file"
+                  accept="image/png, image/jpeg"
+                  onChange={(e: any) => updateFile(e.target.files[0])}
+                />
+              </div>
+            </div>
+          </>
+        }
       />
       <div className="dashboardContainer">
         <div className="cardsContainer">
